+++ backend/src/main/java/com/hans/hans/domain/bodygame/dto/BodyGameStartRequestDto.java
@@ -13,5 +13,5 @@ public class BodyGameStartRequestDto {
     @NotBlank(message = "총 문제수는 필수 입력값입니다.")
     private int totalQuestion;
     @NotBlank(message = "난이도는 필수 입력값입니다.")
-    private int difficulty;
+    private Long difficulty;
 }
